post-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
pydantic==2.7.4
pydantic-settings==2.3.4
httpx==0.27.0
